pyproject.toml
[tool.poetry]
name = "servprog78"
version = "0.1.0"
description = ""
authors = ["ay <artem.a.yats@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
sqlalchemy = "^2.0.30"
fastapi = "^0.111.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
